WIP refactor: move anchoring logic to helper service
