+++ src/lib/popover/popover-anchor.directive.ts
@@ -3,47 +3,30 @@ import {
   ElementRef,
   EventEmitter,
   Input,
-  NgZone,
   OnInit,
   OnDestroy,
-  Optional,
   Output,
   ViewContainerRef
 } from '@angular/core';
-import {
-  ConnectedPositionStrategy,
-  HorizontalConnectionPos,
-  Overlay,
-  OverlayRef,
-  OverlayConfig,
-  ScrollStrategy,
-  VerticalConnectionPos,
-} from '@angular/cdk/overlay';
-import { Direction, Directionality } from '@angular/cdk/bidi';
-import { TemplatePortal } from '@angular/cdk/portal';
 import { ESCAPE } from '@angular/cdk/keycodes';
 import { Subject } from 'rxjs/Subject';
-import { Observable } from 'rxjs/Observable';
 import { take } from 'rxjs/operators/take';
-import { switchMap } from 'rxjs/operators/switchMap';
 import { takeUntil } from 'rxjs/operators/takeUntil';
 import { filter } from 'rxjs/operators/filter';
 import { tap } from 'rxjs/operators/tap';
-import { merge } from 'rxjs/observable/merge';
 
-import {
-  SatPopover,
-  SatPopoverHorizontalAlign,
-  SatPopoverVerticalAlign,
-  SatPopoverScrollStrategy,
-} from './popover.component';
+import { SatPopover } from './popover.component';
 import { NotificationAction, PopoverNotificationService } from './notification.service';
 import { getInvalidPopoverError } from './popover.errors';
+import { PopoverAnchoringService } from './popover-anchoring.service';
 
 @Directive({
   selector: '[satPopoverAnchorFor]',
   exportAs: 'satPopoverAnchor',
-  providers: [PopoverNotificationService],
+  providers: [
+    PopoverNotificationService,
+    PopoverAnchoringService,
+  ],
 })
 export class SatPopoverAnchor implements OnInit, OnDestroy {
 
@@ -71,25 +54,17 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
     return this._popoverOpen;
   }
 
-  /** Reference to the overlay containing the popover component. */
-  _overlayRef: OverlayRef;
-
   /** Whether the popover is presently open. */
   private _popoverOpen = false;
-
-  /** Reference to a template portal where the overlay will be attached. */
-  private _portal: TemplatePortal<any>;
 
   /** Emits when the directive is destroyed. */
   private _onDestroy = new Subject<void>();
 
   constructor(
-    private _overlay: Overlay,
     private _elementRef: ElementRef,
     private _viewContainerRef: ViewContainerRef,
     private _notifications: PopoverNotificationService,
-    private _ngZone: NgZone,
-    @Optional() private _dir: Directionality
+    public _anchoring: PopoverAnchoringService,
   ) { }
 
   ngOnInit() {
@@ -111,8 +86,7 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
   /** Opens the popover. */
   openPopover(): void {
     if (!this._popoverOpen) {
-      this._createOverlay();
-      this._overlayRef.attach(this._portal);
+      this._anchoring.createOverlay(this.attachedPopover, this._viewContainerRef, this._elementRef);
       this._subscribeToBackdrop();
       this._subscribeToEscape();
       this._subscribeToDetachments();
@@ -122,22 +96,17 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
 
   /** Closes the popover. */
   closePopover(value?: any): void {
-    if (this._overlayRef) {
+    if (this._anchoring._overlayRef) {
       this._saveClosedState(value);
-      this._overlayRef.detach();
+      this._anchoring._overlayRef.detach();
     }
-  }
-
-  /** Gets the text direction of the containing app. */
-  private _getDirection(): Direction {
-    return this._dir && this._dir.value === 'rtl' ? 'rtl' : 'ltr';
   }
 
   /** Removes the popover from the DOM. Does NOT update open state. */
   private _destroyPopover(): void {
-    if (this._overlayRef) {
-      this._overlayRef.dispose();
-      this._overlayRef = null;
+    if (this._anchoring._overlayRef) {
+      this._anchoring._overlayRef.dispose();
+      this._anchoring._overlayRef = null;
     }
   }
 
@@ -146,8 +115,8 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
    * has been closed to destroy it.
    */
   private _destroyPopoverOnceClosed(): void {
-    if (this.isPopoverOpen() && this._overlayRef) {
-      this._overlayRef.detachments().pipe(
+    if (this.isPopoverOpen() && this._anchoring._overlayRef) {
+      this._anchoring._overlayRef.detachments().pipe(
         take(1),
         takeUntil(this._onDestroy)
       ).subscribe(() => this._destroyPopover());
@@ -192,7 +161,7 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
 
   /** Close popover when backdrop is clicked. */
   private _subscribeToBackdrop(): void {
-    this._overlayRef
+    this._anchoring._overlayRef
       .backdropClick()
       .pipe(
         tap(() => this.attachedPopover.backdropClicked.emit()),
@@ -205,7 +174,7 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
 
   /** Close popover when escape keydown event occurs. */
   private _subscribeToEscape(): void {
-    this._overlayRef
+    this._anchoring._overlayRef
       .keydownEvents()
       .pipe(
         tap(event => this.attachedPopover.overlayKeydown.emit(event)),
@@ -219,7 +188,7 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
 
   /** Set state back to closed when detached. */
   private _subscribeToDetachments(): void {
-    this._overlayRef
+    this._anchoring._overlayRef
       .detachments()
       .pipe(takeUntil(this._onDestroy))
       .subscribe(() => this._saveClosedState());
@@ -245,214 +214,4 @@ export class SatPopoverAnchor implements OnInit, OnDestroy {
     }
   }
 
-  /** Create an overlay to be attached to the portal. */
-  private _createOverlay(): void {
-    if (!this._overlayRef) {
-      this._portal = new TemplatePortal(this.attachedPopover._templateRef, this._viewContainerRef);
-      const config = this._getOverlayConfig();
-      this._subscribeToPositionChanges(config.positionStrategy as ConnectedPositionStrategy);
-      this._overlayRef = this._overlay.create(config);
-    }
-  }
-
-  /** Create and return a config for creating the overlay. */
-  private _getOverlayConfig(): OverlayConfig {
-    const config = new OverlayConfig({
-      positionStrategy: this._getPositionStrategy(),
-      hasBackdrop: this.attachedPopover.hasBackdrop,
-      backdropClass: this.attachedPopover.backdropClass || 'cdk-overlay-transparent-backdrop',
-      scrollStrategy: this._getScrollStrategyInstance(this.attachedPopover.scrollStrategy),
-      direction: this._getDirection(),
-    });
-
-    return config;
-  }
-
-  /**
-   * Listen to changes in the position of the overlay and set the correct alignment classes,
-   * ensuring that the animation origin is correct, even with a fallback position.
-   */
-  private _subscribeToPositionChanges(position: ConnectedPositionStrategy): void {
-    position.onPositionChange
-      .pipe(takeUntil(this._onDestroy))
-      .subscribe(change => {
-        // Position changes may occur outside the Angular zone
-        this._ngZone.run(() => {
-          this.attachedPopover._setAlignmentClasses(
-            getHorizontalPopoverAlignment(change.connectionPair.overlayX),
-            getVerticalPopoverAlignment(change.connectionPair.overlayY),
-          );
-        });
-      });
-  }
-
-  /** Map a scroll strategy string type to an instance of a scroll strategy. */
-  private _getScrollStrategyInstance(strategy: SatPopoverScrollStrategy): ScrollStrategy {
-    switch (strategy) {
-      case 'block':
-        return this._overlay.scrollStrategies.block();
-      case 'reposition':
-        return this._overlay.scrollStrategies.reposition();
-      case 'close':
-        return this._overlay.scrollStrategies.close();
-      case 'noop':
-      default:
-        return this._overlay.scrollStrategies.noop();
-    }
-  }
-
-  /** Create and return a position strategy based on config provided to the component instance. */
-  private _getPositionStrategy(): ConnectedPositionStrategy {
-    const horizontalTarget = this.attachedPopover.horizontalAlign;
-    const verticalTarget = this.attachedPopover.verticalAlign;
-
-    // Attach the overlay at the preferred position
-    const {originX, overlayX} = getHorizontalConnectionPosPair(horizontalTarget);
-    const {originY, overlayY} = getVerticalConnectionPosPair(verticalTarget);
-    const strategy = this._overlay.position()
-      .connectedTo(this._elementRef, {originX, originY}, {overlayX, overlayY})
-      .withDirection(this._getDirection());
-
-    // Add fallbacks based on the preferred positions
-    this._addFallbacks(strategy, horizontalTarget, verticalTarget);
-
-    return strategy;
-  }
-
-  /** Add fallbacks to a given strategy based around target alignments. */
-  private _addFallbacks(strategy: ConnectedPositionStrategy, hTarget: SatPopoverHorizontalAlign,
-        vTarget: SatPopoverVerticalAlign): void {
-    // Determine if the target alignments overlap the anchor
-    const horizontalOverlapAllowed = hTarget !== 'before' && hTarget !== 'after';
-    const verticalOverlapAllowed = vTarget !== 'above' && vTarget !== 'below';
-
-    // If a target alignment doesn't cover the anchor, don't let any of the fallback alignments
-    // cover the anchor
-    const possibleHorizontalAlignments = horizontalOverlapAllowed ?
-      ['before', 'start', 'center', 'end', 'after'] :
-      ['before', 'after'];
-    const possibleVerticalAlignments = verticalOverlapAllowed ?
-      ['above', 'start', 'center', 'end', 'below'] :
-      ['above', 'below'];
-
-    // Create fallbacks for each allowed prioritized fallback alignment combo
-    const fallbacks = [];
-    prioritizeAroundTarget(hTarget, possibleHorizontalAlignments).forEach(h => {
-      prioritizeAroundTarget(vTarget, possibleVerticalAlignments).forEach(v => {
-        fallbacks.push({h, v});
-      });
-    });
-
-    // Remove the first fallback since it will be the target alignment that is already applied
-    fallbacks.slice(1, fallbacks.length)
-      .forEach(({h, v}) => this._applyFallback(strategy, h, v));
-  }
-
-  /**
-   * Convert a specific horizontal and vertical alignment into a fallback and apply it to
-   * the strategy.
-   */
-  private _applyFallback(strategy, horizontalAlign, verticalAlign): void {
-    const {originX, overlayX} = getHorizontalConnectionPosPair(horizontalAlign);
-    const {originY, overlayY} = getVerticalConnectionPosPair(verticalAlign);
-    strategy.withFallbackPosition({originX, originY}, {overlayX, overlayY});
-  }
-
-}
-
-/** Helper function to convert alignment to origin/overlay position pair. */
-function getHorizontalConnectionPosPair(h: SatPopoverHorizontalAlign):
-    {originX: HorizontalConnectionPos, overlayX: HorizontalConnectionPos} {
-  switch (h) {
-    case 'before':
-      return {originX: 'start', overlayX: 'end'};
-    case 'start':
-      return {originX: 'start', overlayX: 'start'};
-    case 'end':
-      return {originX: 'end', overlayX: 'end'};
-    case 'after':
-      return {originX: 'end', overlayX: 'start'};
-    default:
-      return {originX: 'center', overlayX: 'center'};
-  }
-}
-
-/** Helper function to convert alignment to origin/overlay position pair. */
-function getVerticalConnectionPosPair(v: SatPopoverVerticalAlign):
-    {originY: VerticalConnectionPos, overlayY: VerticalConnectionPos} {
-  switch (v) {
-    case 'above':
-      return {originY: 'top', overlayY: 'bottom'};
-    case 'start':
-      return {originY: 'top', overlayY: 'top'};
-    case 'end':
-      return {originY: 'bottom', overlayY: 'bottom'};
-    case 'below':
-      return {originY: 'bottom', overlayY: 'top'};
-    default:
-      return {originY: 'center', overlayY: 'center'};
-  }
-}
-
-/** Helper function to convert an overlay connection position to equivalent popover alignment. */
-function getHorizontalPopoverAlignment(h: HorizontalConnectionPos): SatPopoverHorizontalAlign {
-  if (h === 'start') {
-    return 'after';
-  }
-
-  if (h === 'end') {
-    return 'before';
-  }
-
-  return 'center';
-}
-
-/** Helper function to convert an overlay connection position to equivalent popover alignment. */
-function getVerticalPopoverAlignment(v: VerticalConnectionPos): SatPopoverVerticalAlign {
-  if (v === 'top') {
-    return 'below';
-  }
-
-  if (v === 'bottom') {
-    return 'above';
-  }
-
-  return 'center';
-}
-
-/**
- * Helper function that takes an ordered array options and returns a reorderded
- * array around the target item. e.g.:
- *
- * target: 3; options: [1, 2, 3, 4, 5, 6, 7];
- *
- * return: [3, 4, 2, 5, 1, 6, 7]
- */
-function prioritizeAroundTarget<T>(target: T, options: T[]): T[] {
-  const targetIndex = options.indexOf(target);
-
-  // Set the first item to be the target
-  const reordered = [target];
-
-  // Make left and right stacks where the highest priority item is last
-  const left = options.slice(0, targetIndex);
-  const right = options.slice(targetIndex + 1, options.length).reverse();
-
-  // Alternate between stacks until one is empty
-  while (left.length && right.length) {
-    reordered.push(right.pop());
-    reordered.push(left.pop());
-  }
-
-  // Flush out right side
-  while (right.length) {
-    reordered.push(right.pop());
-  }
-
-  // Flush out left side
-  while (left.length) {
-    reordered.push(left.pop());
-  }
-
-  return reordered;
 }
